Clean module root by moving all unnecessary files to corresponding dir's
